backend/pyproject.toml
[project]
name = "alumniapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "beanie>=1.29.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.13",
    "uvicorn>=0.34.3",
]

[tool.ruff]
line-length = 100
fix = true
select = ["ALL"]
ignore = ["ANN", "D", "E712"]
